pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ffs-rag"
version = "0.1.0"
description = "Hands-on RAG course for experienced Python developers"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]

dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "numpy>=1.26",
  "scipy>=1.13",
  "scikit-learn>=1.5",
  "qdrant-client[fastembed]>=1.14.2",
  "rank-bm25>=0.2.2",
  "redis>=5.0",
  "prometheus-client>=0.20",
  "opentelemetry-sdk>=1.25.0",
  "opentelemetry-exporter-otlp>=1.25.0",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "opentelemetry-instrumentation-logging>=0.46b0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "orjson>=3.10",
  "tenacity>=8.4",
  "starlette>=0.47.3",
  "pytest>=8.3",
]

[project.optional-dependencies]
dev = [
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.5",
  "black>=24.8",
  "mypy>=1.10",
  "types-redis",
  "types-requests",
  "pre-commit>=3.7",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B", "UP", "N", "S", "RUF"]
lint.ignore = ["E501", "B008", "RUF001", "S324", "S603"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
plugins = []
explicit_package_bases = true
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
